fix for mac

+++ mutex.c
@@ -32,22 +32,23 @@ void mutex_lock_do (Mutex *mutex, const char *func)
     ASSERT (mutex->initialized, "Error in mutex_lock_do called from %s: mutex not initialized", func);
 
     int ret = pthread_mutex_lock (&mutex->mutex); 
-    ASSERT (!ret, "Error in mutex_lock_do called from %s by %"PRIu64": pthread_mutex_lock failed: %s", func, pthread_self(), strerror (ret)); 
+    ASSERT (!ret, "Error in mutex_lock_do called from %s by %"PRIu64": pthread_mutex_lock failed: %s", 
+            func, (uint64_t)pthread_self(), strerror (ret)); 
 
     mutex->lock_func = func; // mutex->lock_func is protected by the mutex
 
     if (flag.show_mutex && !strncmp (mutex->name, flag.show_mutex, 8)) // only 8 chars so we can catch all genome_muteces[%u]
-        fprintf (stderr, "LOCK  : Mutex %s by thread %"PRIu64"\n", mutex->name, pthread_self());
+        fprintf (stderr, "LOCK  : Mutex %s by thread %"PRIu64"\n", mutex->name, (uint64_t)pthread_self());
 }
 
 void mutex_unlock_do (Mutex *mutex, const char *func, uint32_t line) 
 { 
     ASSERT (mutex->initialized, "Error in mutex_lock_do called from %s:%u mutex not initialized", func, line);
     ASSERT (mutex->lock_func, "Error in mutex_unlock_do called from %s:%u by thread=%"PRIu64": mutex %s is not locked", 
-            func, line, pthread_self(), mutex->name);
+            func, line, (uint64_t)pthread_self(), mutex->name);
 
     if (flag.show_mutex && !strncmp (mutex->name, flag.show_mutex, 8))
-        fprintf (stderr, "UNLOCK: Mutex %s by thread %"PRIu64"\n", mutex->name, pthread_self());
+        fprintf (stderr, "UNLOCK: Mutex %s by thread %"PRIu64"\n", mutex->name, (uint64_t)pthread_self());
 
     mutex->lock_func = NULL; // mutex->lock_func is protected by the mutex
